requirements.txt
crewai>=0.28.0
crewai-tools>=0.2.0
openai>=1.12.0
streamlit>=1.31.0
python-dotenv>=1.0.0
mcp>=0.9.0
pydantic>=2.6.0
langchain-openai>=0.1.0
notion-client>=2.0.0
nest-asyncio>=1.5.0

